requirements.txt
requests~=2.28.2
aiogram~=2.25.1
python_dotenv
python-dotenv~=1.0.0